refactor(library structure): Refactor structure for an easier access and export/import.
