+++ src/ui/Anchor/index.js
@@ -1,0 +1,1 @@
+export { default as Anchor } from './Anchor.atom'
