+++ src/ui/Header/index.js
@@ -1,0 +1,1 @@
+export { default as Header } from './Header.molecule'
