+++ src/ui/Input/index.js
@@ -1,0 +1,1 @@
+export { default as Input } from './Input.atom'
