+++ src/ui/Modal/index.js
@@ -1,0 +1,1 @@
+export { default as Modal } from './Modal.atom'
