+++ src/ui/Spinner/Spinner.examples.md
